+++ src/main/java/kr/hhplus/be/server/ServerApplication.java
@@ -2,7 +2,6 @@ package kr.hhplus.be.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 public class ServerApplication {
