+++ src/main/java/kr/hhplus/be/server/coupon/infra/repositpry/CouponTypeJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.coupon.infra;
+package kr.hhplus.be.server.coupon.infra.repositpry;
 
 import kr.hhplus.be.server.coupon.domain.entity.CouponType;
 import org.springframework.data.jpa.repository.JpaRepository;
