+++ src/main/java/kr/hhplus/be/server/coupon/infra/repositpry/adapter/CouponRepositoryAdapter.java
@@ -1,7 +1,8 @@
-package kr.hhplus.be.server.coupon.infra;
+package kr.hhplus.be.server.coupon.infra.repositpry.adapter;
 
 import kr.hhplus.be.server.coupon.domain.entity.Coupon;
-import kr.hhplus.be.server.coupon.domain.repository.CouponRepository;
+import kr.hhplus.be.server.coupon.infra.repositpry.port.CouponRepository;
+import kr.hhplus.be.server.coupon.infra.repositpry.CouponJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
