+++ src/main/java/kr/hhplus/be/server/coupon/infra/repositpry/port/CouponRepository.java
@@ -1,11 +1,9 @@
-package kr.hhplus.be.server.coupon.infra;
+package kr.hhplus.be.server.coupon.infra.repositpry.port;
 
 import kr.hhplus.be.server.coupon.domain.entity.Coupon;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface CouponJpaRepository extends JpaRepository<Coupon, Long> {
+public interface CouponRepository {
     Coupon save(Coupon coupon);
     Optional<Coupon> findByUserIdAndCouponTypeId(Long userId, Long couponTypeId);
 }
