--- src/main/java/kr/hhplus/be/server/product/infra/ProductRepositoryAdapter.java
@@ -1,19 +1,0 @@
-package kr.hhplus.be.server.product.infra;
-
-import kr.hhplus.be.server.product.domain.entity.Product;
-import kr.hhplus.be.server.product.domain.repository.ProductRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-import java.util.Optional;
-
-@Component
-@RequiredArgsConstructor
-public class ProductRepositoryAdapter implements ProductRepository {
-    private final ProductJpaRepository jpaRepository;
-
-    @Override
-    public Optional<Product> findById(Long id) {
-        return Optional.empty();
-    }
-}
