+++ src/main/java/kr/hhplus/be/server/product/infra/repository/ProductJpaRepository.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.product.infra.repository;
+
+import kr.hhplus.be.server.mock.model.Coupon;
+import kr.hhplus.be.server.product.domain.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductJpaRepository extends JpaRepository<Product, Long> {
+
+}
