+++ src/main/java/kr/hhplus/be/server/user/infra/reposistory/BalanceHistoryJpaRepository.java
@@ -1,7 +1,6 @@
-package kr.hhplus.be.server.user.infra;
+package kr.hhplus.be.server.user.infra.reposistory;
 
 import kr.hhplus.be.server.user.domain.entity.UserBalanceHistory;
-import kr.hhplus.be.server.user.domain.repository.UserRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BalanceHistoryJpaRepository extends JpaRepository<UserBalanceHistory, Long> {
