+++ src/main/java/kr/hhplus/be/server/user/infra/reposistory/UserJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.user.infra;
+package kr.hhplus.be.server.user.infra.reposistory;
 
 import kr.hhplus.be.server.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
