+++ src/main/java/kr/hhplus/be/server/user/infra/reposistory/adapter/BalanceHistoryAdapter.java
@@ -1,7 +1,8 @@
-package kr.hhplus.be.server.user.infra;
+package kr.hhplus.be.server.user.infra.reposistory.adapter;
 
 import kr.hhplus.be.server.user.domain.entity.UserBalanceHistory;
-import kr.hhplus.be.server.user.domain.repository.BalanceHistoryRepository;
+import kr.hhplus.be.server.user.infra.reposistory.BalanceHistoryJpaRepository;
+import kr.hhplus.be.server.user.infra.reposistory.port.BalanceHistoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -10,7 +11,6 @@ import org.springframework.stereotype.Component;
 public class BalanceHistoryAdapter implements BalanceHistoryRepository {
 
     private final BalanceHistoryJpaRepository jpaRepository;
-
 
     @Override
     public UserBalanceHistory save(UserBalanceHistory userBalanceHistory) {
