+++ src/main/java/kr/hhplus/be/server/user/infra/reposistory/port/BalanceHistoryRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.user.domain.repository;
+package kr.hhplus.be.server.user.infra.reposistory.port;
 
 import kr.hhplus.be.server.user.domain.entity.UserBalanceHistory;
