+++ src/main/java/kr/hhplus/be/server/user/infra/reposistory/port/UserRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.user.domain.repository;
+package kr.hhplus.be.server.user.infra.reposistory.port;
 
 import kr.hhplus.be.server.user.domain.entity.User;
 
@@ -6,6 +6,5 @@ import java.util.Optional;
 
 public interface UserRepository {
     Optional<User> findById(Long userId);
-
     void save(User user);
 }
